core-app/requirements.txt
streamlit==1.29.0
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2
tensorflow==2.15.0
ta==0.10.2
ccxt==4.1.65
python-binance==1.0.19
websocket-client==1.6.4
openai==1.3.7
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
plotly==5.17.0
loguru==0.7.2
python-dotenv==1.0.0
requests==2.31.0
jinja2==3.1.2
pytest==7.4.3
black==23.11.0
ruff==0.1.6
pydantic==2.5.0
asyncio-mqtt==0.16.1
schedule==1.2.0